Download GML geodata file for Prague from CUZK

Download GML geodata file for Prague from CUZK

get_cuzk_prague(
  date = strftime(lubridate::rollback(Sys.Date()), format = "%Y%m%d"),
  path = ".",
  cachetime = 32
)

Arguments

date

Date of file, must be last day of month. Defaults to end of previous month.

path

Where to save the downloaded zip file and unzipped xml file

cachetime

Redownload if the zip file found in path exists and is older than cachetime. Defaults to 32. Set to 0 to always redownload. Only checks existence and timestamp of downloaded zip file, not unzipped file.

Value

string Path to downloaded and unzipped file incl. path