Prague base maps for ggplot2

Include raster maps from Prague geoportal in your ggplot map

prg_basemap(
  data,
  image_service = "orto",
  layer = "",
  width = 900,
  alpha = 1,
  buffer = 0,
  verbose = F
)

Arguments

data

sf data frame from which to extract the bounding box

image_service

map service from which to draw the map; prg_endpoints provides details.

layer

layer from map service to use, see https://mpp.praha.eu/arcgis/rest/services/

width

width in pixels, in effect sets image resolution; integer or "max"

alpha

transparency of the tile

buffer

distance between feature end and image end; for EPSG 5514 in meters.

verbose

display information on image URLs and image processing.

Value

list including raster annotation layers for ggplot2

See also

Examples

if (FALSE) {
praha1 <- CzechData::load_RUIAN_settlement(prg_kod, "MOMC_P", WGS84 = F) %>%
  filter(nazev == 'Praha 1')

ggplot() +
  prg_basemap(data = praha1, alpha = .8, buffer = 200,
              image_service = 'mapy_archiv', layer = 6) +
  geom_sf(data = praha1, fill = alpha("red", 0.6), colour = NA) +
  theme_void()
}